# Update to Gulp4 and use Terser instead of UglifyJS for JS minification