--- a/package.json
+++ b/package.json
@@ -8,7 +8,7 @@
   "author": "Stephen Murray <spmurrayzzz>",
   "license": "MIT",
   "devDependencies": {
-    "gulp": "3.9.1",
+    "gulp": "4.0.2",
     "gulp-buffer": "0.0.2",
     "gulp-clean-css": "2.0.12",
     "gulp-concat-css": "2.3.0",
@@ -17,12 +17,13 @@
     "gulp-plumber": "1.1.0",
     "gulp-rename": "1.2.2",
     "gulp-sourcemaps": "1.6.0",
-    "gulp-uglify": "1.5.4",
-    "gulp-util": "^3.0.7",
+    "gulp-terser": "1.3.2",
+    "gulp-util": "3.0.7",
     "gulp-zip": "3.2.0",
     "handlebars": "4.0.5",
     "rollup-stream": "1.11.0",
-    "vinyl-buffer": "1.0.0",
-    "vinyl-source-stream": "1.1.0"
+    "vinyl": "2.2.0",
+    "vinyl-buffer": "1.0.1",
+    "vinyl-source-stream": "2.0.0"
   }
 }