--- a/tasks/watch.js
+++ b/tasks/watch.js
@@ -3,12 +3,12 @@
 const gulp = require('gulp');
 const livereload  = require('gulp-livereload');
 
-module.exports = () => {
-  gulp.task( 'watch', () => {
-    livereload.listen();
-    return gulp.watch(
-      [ 'src/js/**/*.js', 'src/css/**/*.css', 'src/index.hbs' ],
-      [ 'build', 'css', 'template', 'zip', 'report' ]
-    );
-  });
-};
+function watch() {
+  livereload.listen();
+  return gulp.watch(
+    [ 'src/js/**/*.js', 'src/css/**/*.css', 'src/index.hbs' ],
+    [ 'build', 'css', 'template', 'zip', 'report' ]
+  );
+}
+
+exports.watch = watch;