--- a/tasks/watch.js
+++ b/tasks/watch.js
@@ -2,12 +2,15 @@
 
 const gulp = require('gulp');
 const livereload  = require('gulp-livereload');
+const distTasks = require('./dist.js');
 
 function watch() {
   livereload.listen();
   return gulp.watch(
     [ 'src/js/**/*.js', 'src/css/**/*.css', 'src/index.hbs' ],
-    [ 'build', 'css', 'template', 'zip', 'report' ]
+    // Gulp4 requires a function to be passed to watch()
+    // a call to series() with one task returns a function
+    gulp.series(distTasks.dist)
   );
 }