Updating to Gulp4 allows the project to be compatible with Node versions >10.0.0.

Performance is also improved (possibly because of the switch to Gulp4?). Results taken from empty project:
    `build-full`: 168ms -> 63 ms
    `css`: 103ms -> 33ms
    `report`: 49ms -> 7.94ms

Using Terser [usually gives a smaller zip file size than UglifyJS](https://blog.logrocket.com/uglify-vs-babel-minify-vs-terser-a-mini-battle-royale/). In this case, the final zip file size was reduced from 639 bytes to 628 bytes.